started functionality to add new users to chat

+++ frontend/src/components/TheChatSettings.vue
@@ -35,9 +35,38 @@
                 </template>
                  <v-divider/>
               </v-list>
+              <div v-if="this.userToAddVisible">
+                <p class="subheading">Teilnehmer hinzufügen:</p>
+                <v-list class="user-to-add-list" two-line>
+                  <template v-for="(user, index) in usersNotInChat">
+                    <v-divider v-if="index!=0"  :key="index"></v-divider>
+
+                    <v-list-tile :key="user.username" avatar>
+                      <v-list-tile-avatar>
+                        <img
+                          :src="user.picture"
+                          @error="user.picture=`${publicPath}images/user_icon.svg`"
+                        >
+                      </v-list-tile-avatar>
+
+                      <v-list-tile-content>
+                        <v-list-tile-title v-html="user.username"></v-list-tile-title>
+                        <v-list-tile-sub-title v-html="user.status"></v-list-tile-sub-title>
+                      </v-list-tile-content>
+                      <v-btn flat fab @click="deleteUserFromChat(user.id)">
+                        <v-icon>add</v-icon>
+                      </v-btn>
+                    </v-list-tile>
+                  </template>
+                  <v-divider/>
+                </v-list>
+              </div>
               <div class="--center-content">
-              <v-btn round class="add-btn" color="primary">
+              <v-btn round class="add-btn" color="primary" @click="userToAddVisible=true" v-if="!userToAddVisible">
                   <v-icon class="__icon">add</v-icon>Neuer Teilnehmer
+                </v-btn>
+                <v-btn round class="add-btn" color="primary" @click="userToAddVisible=false" v-if="userToAddVisible">
+                  Abbrechen
                 </v-btn>
               </div>
             </div>
@@ -67,6 +96,8 @@ export default class ChatSettings extends Vue {
   private status: string = 'Kein Status';
   private user?: User;
   private usersInChat: User[] = [];
+  private usersNotInChat: User[] = [];
+  private userToAddVisible: boolean = false;
 
   @Prop()
   private chat: Chat = {id: undefined, chatname: 'Kein Name', picture: '', users: []};
@@ -74,6 +105,7 @@ export default class ChatSettings extends Vue {
   private mounted() {
     this.user = this.$store.state.loggedInUser;
     this.setUsersInChat();
+    this.setUsersNotInChat();
   }
 
   private async getUser(userId: number): Promise<User> {
@@ -106,9 +138,19 @@ export default class ChatSettings extends Vue {
       }
   }
 
+  @Watch('chat')
+  private async setUsersNotInChat() {
+      if (this.chat.id !== undefined) {
+        const allUsers = await UserService.getAllUsers();
+        this.usersNotInChat = allUsers.filter(u => !this.usersInChat.includes(u));
+      }
+  }
+
   private async deleteUserFromChat(userId: number) {
-    await ChatService.deleteUserFromChat(userId);
-    this.usersInChat = this.usersInChat.filter(u => u.id !== userId);
+    if (this.usersInChat.length < 1) {
+      await ChatService.deleteUserFromChat(userId);
+      this.usersInChat = this.usersInChat.filter(u => u.id !== userId);
+    }
   }
 }
 </script>
